pyproject.toml
[build-system]
build-backend = "hatchling.build"

requires = [ "hatchling" ]

[project]
name = "openhands-server"
version = "0.1.0"
description = "OpenHands Server - REST/WebSocket interface for OpenHands AI Agent"
readme = "README.md"
license = { text = "MIT" }
authors = [ { name = "OpenHands Team", email = "contact@all-hands.dev" } ]
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "openhands-sdk @ git+https://github.com/All-Hands-AI/agent-sdk.git@b3ddfb3468c22a37fbb87d96323914d23bacb4ed#subdirectory=openhands/sdk",
  "openhands-tools @ git+https://github.com/All-Hands-AI/agent-sdk.git@b3ddfb3468c22a37fbb87d96323914d23bacb4ed#subdirectory=openhands/tools",
  "fastapi>=0.104.0",
  "uvicorn>=0.31.1",
  "websockets>=12.0",
  "pydantic>=2.0.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "docker (>=7.1.0,<8.0.0)",
]

optional-dependencies.dev = [
  "black>=23",
  "flake8>=6",
  "isort>=5",
  "mypy>=1",
  "pre-commit>=3.7",
  "pytest>=7",
  "ruff>=0.11.8",
  "httpx>=0.25.0",
  "pytest-asyncio>=0.21.0",
]
scripts.openhands-server = "openhands_server.main:main"

[dependency-groups]
dev = [
  "pre-commit>=4.3",
  "pyinstaller>=6.15",
  "pytest>=8.4.1",
  "httpx>=0.25.0",
  "pytest-asyncio>=0.21.0",
]

[tool.poetry]
name = "openhands-server"
version = "0.1.0"
description = "OpenHands Server - REST/WebSocket interface for OpenHands AI Agent"
authors = [ "OpenHands Team <contact@all-hands.dev>" ]
license = "MIT"
readme = "README.md"
packages = [ { include = "openhands_server" } ]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.104.0"
uvicorn = ">=0.31.1"
websockets = ">=12.0"
pydantic = ">=2.0.0"
sqlalchemy = ">=2.0.0"
alembic = ">=1.13.0"
openhands-sdk = { git = "https://github.com/All-Hands-AI/agent-sdk.git", rev = "b3ddfb3468c22a37fbb87d96323914d23bacb4ed", subdirectory = "openhands/sdk" }
openhands-tools = { git = "https://github.com/All-Hands-AI/agent-sdk.git", rev = "b3ddfb3468c22a37fbb87d96323914d23bacb4ed", subdirectory = "openhands/tools" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pre-commit = "^3.7.0"
ruff = "^0.11.8"
httpx = ">=0.28.1"
pytest-asyncio = "^0.21.0"

[tool.poetry.scripts]
openhands-server = "openhands_server.main:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [ "openhands_server" ]

# Keep Poetry configuration for compatibility

[tool.black]
line-length = 88
target-version = [ 'py312' ]

[tool.ruff]
target-version = "py312"

line-length = 88
format.indent-style = "space"
format.quote-style = "double"
format.line-ending = "auto"
format.skip-magic-trailing-comma = false
lint.select = [
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "UP", # pyupgrade
  "W",  # pycodestyle warnings
]
lint.ignore = [
  "B008", # do not perform function calls in argument defaults
  "C901", # too complex
  "E501", # line too long, handled by black
]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.uv.sources]
openhands-sdk = { git = "https://github.com/All-Hands-AI/agent-sdk.git", subdirectory = "openhands/sdk", rev = "b3ddfb3468c22a37fbb87d96323914d23bacb4ed" }